Fix up test
