+++ injected/unit-test/utils.js
@@ -5,6 +5,7 @@ import {
     processConfig,
     satisfiesMinVersion,
     getTabHostname,
+    processAttr,
 } from '../src/utils.js';
 import { polyfillProcessGlobals } from './helpers/polyfill-process-globals.js';
 
@@ -301,6 +302,331 @@ describe('Helpers checks', () => {
             const hostname5 = getTabHostname();
             expect(hostname5).toEqual('example.com');
             reset2();
+        });
+    });
+
+    describe('processAttr', () => {
+        describe('Basic types', () => {
+            it('returns default value when configSetting is undefined', () => {
+                expect(processAttr(/** @type {any} */ (undefined), 'default')).toBe('default');
+            });
+
+            it('returns default value when configSetting is not an object', () => {
+                expect(processAttr(/** @type {any} */ ('string'), 'default')).toBe('default');
+                expect(processAttr(/** @type {any} */ (123), 'default')).toBe('default');
+                expect(processAttr(/** @type {any} */ (true), 'default')).toBe('default');
+            });
+
+            it('returns default value when configSetting has no type', () => {
+                expect(processAttr(/** @type {any} */ ({}), 'default')).toBe('default');
+            });
+
+            it('handles undefined type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'undefined' };
+                expect(processAttr(configSetting)).toBe(undefined);
+            });
+
+            it('handles string type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'string', value: 'hello' };
+                expect(processAttr(configSetting)).toBe('hello');
+            });
+
+            it('handles number type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'number', value: 42 };
+                expect(processAttr(configSetting)).toBe(42);
+            });
+
+            it('handles boolean type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'boolean', value: true };
+                expect(processAttr(configSetting)).toBe(true);
+            });
+
+            it('handles null type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'null', value: null };
+                expect(processAttr(configSetting)).toBe(null);
+            });
+
+            it('handles array type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'array', value: [1, 2, 3] };
+                expect(processAttr(configSetting)).toEqual([1, 2, 3]);
+            });
+
+            it('handles object type', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'object', value: { key: 'value' } };
+                expect(processAttr(configSetting)).toEqual({ key: 'value' });
+            });
+        });
+
+        describe('Function type', () => {
+            it('handles function type with functionName', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = { type: 'function', functionName: 'noop' };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toBe(undefined); // noop returns undefined
+            });
+
+            it('handles function type with functionValue', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'number',
+                        value: 1,
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toBe(1);
+            });
+
+            it('handles function type with complex functionValue', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'string',
+                        value: 'hello world',
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toBe('hello world');
+            });
+
+            it('handles function type with nested object functionValue', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'object',
+                        value: { message: 'test', count: 5 },
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toEqual({ message: 'test', count: 5 });
+            });
+
+            it('handles function type with array functionValue', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'array',
+                        value: ['a', 'b', 'c'],
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toEqual(['a', 'b', 'c']);
+            });
+
+            it('prefers functionName over functionValue when both are present', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionName: 'noop',
+                    functionValue: {
+                        type: 'string',
+                        value: 'should not be used',
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                expect(result()).toBe(undefined); // noop returns undefined
+            });
+        });
+
+        describe('Async support', () => {
+            it('handles async string', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'string',
+                    value: 'boop',
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toBe('boop');
+            });
+
+            it('handles async number', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'number',
+                    value: 123,
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toBe(123);
+            });
+
+            it('handles async boolean', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'boolean',
+                    value: false,
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toBe(false);
+            });
+
+            it('handles async array', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'array',
+                    value: [1, 2, 'test'],
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toEqual([1, 2, 'test']);
+            });
+
+            it('handles async object', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'object',
+                    value: { key: 'async-value', nested: { prop: 'test' } },
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toEqual({ key: 'async-value', nested: { prop: 'test' } });
+            });
+
+            it('handles async null', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'null',
+                    value: null,
+                    async: true,
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toBe(null);
+            });
+
+            it('non-async values should not be wrapped in Promise', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'string',
+                    value: 'not async',
+                };
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(false);
+                expect(result).toBe('not async');
+            });
+        });
+
+        describe('Criteria support', () => {
+            it('handles array with criteria selection - fallback case', () => {
+                /** @type {import('../src/utils.js').ConfigSetting[]} */
+                const configSetting = [
+                    {
+                        type: 'string',
+                        value: 'fallback',
+                    },
+                    {
+                        type: 'string',
+                        value: 'criteria-based',
+                        criteria: { arch: 'SomeOtherArch' }, // This won't match, so should use fallback
+                    },
+                ];
+                const result = processAttr(configSetting);
+                expect(result).toBe('fallback');
+            });
+
+            it('handles async array with criteria selection - fallback case', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting[]} */
+                const configSetting = [
+                    {
+                        type: 'string',
+                        value: 'fallback',
+                        async: true,
+                    },
+                    {
+                        type: 'string',
+                        value: 'criteria-based',
+                        criteria: { arch: 'SomeOtherArch' }, // This won't match, so should use fallback
+                        async: true,
+                    },
+                ];
+                const result = processAttr(configSetting);
+                expect(result instanceof Promise).toBe(true);
+                const resolvedValue = await result;
+                expect(resolvedValue).toBe('fallback');
+            });
+        });
+
+        describe('Complex combinations', () => {
+            it('handles function with async functionValue', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'string',
+                        value: 'async result',
+                        async: true,
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                const functionResult = result();
+                expect(functionResult instanceof Promise).toBe(true);
+            });
+
+            it('function returns correct async value', async () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'object',
+                        value: { data: 'complex async object' },
+                        async: true,
+                    },
+                };
+                const result = processAttr(configSetting);
+                const functionResult = result();
+                const resolvedValue = await functionResult;
+                expect(resolvedValue).toEqual({ data: 'complex async object' });
+            });
+
+            it('nested function with nested processAttr calls', () => {
+                /** @type {import('../src/utils.js').ConfigSetting} */
+                const configSetting = {
+                    type: 'function',
+                    functionValue: {
+                        type: 'function',
+                        functionValue: {
+                            type: 'number',
+                            value: 42,
+                        },
+                    },
+                };
+                const result = processAttr(configSetting);
+                expect(typeof result).toBe('function');
+                const nestedFunction = result();
+                expect(typeof nestedFunction).toBe('function');
+                expect(nestedFunction()).toBe(42);
+            });
         });
     });
 });
